do incoming different

+++ bin/dvd-to-vob.sh
@@ -14,7 +14,11 @@ DEST_D=${1:-vobs}
 mount "$DEVNAME"
 
 SRC_D=$(df "$DEVNAME" | tail -1 | awk '{ printf $6 }')
-[ -z "$SRC_D" ] && exit 10
+if [ -z "$SRC_D" ]; then
+    echo "ERROR: Mounting error for $DEVNAME"
+    eject "$DEVNAME"
+    exit 10
+fi
 
 # set all the env variables here. (happens when not entering through udev)
 if [ -z "$ID_FS_LABEL" ] || [ -z "$ID_FS_UUID" ]; then
@@ -23,26 +27,37 @@ fi
 
 [ -z "$DVD_NAME" ] && DVD_NAME="$ID_FS_LABEL"
 [ -z "$DVD_NAME" ] || [ "$DVD_NAME" = "DVD_VIDEO" ] && DVD_NAME="$ID_FS_UUID"
-[ -z "$DVD_NAME" ] && (echo "No DVD_NAME"; exit 11)
+if [ -z "$DVD_NAME" ]; then
+    echo "ERROR: No DVD_NAME"
+    eject "$DEVNAME"
+    exit 11
+fi
+
 echo "DVD_NAME=$DVD_NAME"
 
-if [ -n "$DVD_NAME" ] && [ -d "$DEST_D/$DVD_NAME" ]; then
-    echo "Movie '$DVD_NAME' already exists"
+if [ -d "$DEST_D/$DVD_NAME" ]; then
+    echo "Movie '$DVD_NAME' already exists. Nothing to do"
     exit 0
+elif [ -d "$DEST_D/.$DVD_NAME]" ]; then
+    echo "WARNING: Cleaning up previous workdir..."
+    rm -rf "$DEST_D/.$DVD_NAME"
 fi
 
-mkdir -p "$DEST_D"
-touch "$DEST_D/$DVD_NAME.incoming"
+mkdir -p "$DEST_D/.$DVD_NAME"
 
 # TODO: timeout in case the copy gets stuck
-mkdir -p "$DEST_D/$DVD_NAME"
-if vobcopy -M -i "$SRC_D" -o "$DEST_D/$DVD_NAME" -t "$DVD_NAME"; then
+echo "Starting copy to $DEST_D..."
+if vobcopy \
+    --large-file \
+    --input-dir "$SRC_D" \
+    -o "$DEST_D/.$DVD_NAME" \
+    -t "$DVD_NAME"
+then
     echo "SUCCESS"
+    mv "$DEST_D/.$DVD_NAME" "$DEST_D/$DVD_NAME"
 else
     echo "FAILED with exit code $?"
-    rm -rf "$DEST_D/$DVD_NAME"
+    rm -rf "${DEST_D:?}/.$DVD_NAME"
 fi
-
-rm "$DEST_D/$DVD_NAME.incoming"
 
 eject "$DEVNAME"
